+++ backend/src/model/entity.rs
@@ -2,6 +2,7 @@
 
 pub mod layer_impl;
 pub mod map_impl;
+pub mod plantings;
 pub mod plantings_impl;
 pub mod plants_impl;
 pub mod seed_impl;
@@ -9,12 +10,11 @@ pub mod seed_impl;
 use chrono::NaiveDate;
 use chrono::NaiveDateTime;
 
-use diesel::AsChangeset;
 use diesel::QueryableByName;
 use diesel::{Identifiable, Insertable, Queryable};
 use postgis_diesel::types::Point;
 
-use crate::schema::{layers, maps, plantings, plants, seeds};
+use crate::schema::{layers, maps, plants, seeds};
 
 use super::r#enum::privacy_options::PrivacyOptions;
 use super::r#enum::{
@@ -786,74 +786,4 @@ pub struct NewLayer {
     pub name: String,
     /// A flag indicating if this layer is an user created alternative.
     pub is_alternative: bool,
-}
-
-/// The `Planting` entity.
-#[derive(Debug, Clone, Identifiable, Queryable)]
-#[diesel(table_name = plantings)]
-pub struct Planting {
-    /// The database id of the record.
-    pub id: i32,
-    /// The plant that is planted.
-    pub layer_id: i32,
-    /// The plant that is planted.
-    pub plant_id: i32,
-    /// The x coordinate of the position on the map.
-    pub x: i32,
-    /// The y coordinate of the position on the map.
-    pub y: i32,
-    /// The width of the plant on the map.
-    pub width: i32,
-    /// The height of the plant on the map.
-    pub height: i32,
-    /// The rotation in degrees (0-360) of the plant on the map.
-    pub rotation: f32,
-    /// The x scale of the plant on the map.
-    pub scale_x: f32,
-    /// The y scale of the plant on the map.
-    pub scale_y: f32,
-}
-
-/// The `Planting` entity.
-#[derive(Debug, Clone, Insertable)]
-#[diesel(table_name = plantings)]
-pub struct NewPlanting {
-    /// The plant that is planted.
-    pub layer_id: i32,
-    /// The plant that is planted.
-    pub plant_id: i32,
-    /// The x coordinate of the position on the map.
-    pub x: i32,
-    /// The y coordinate of the position on the map.
-    pub y: i32,
-    /// The width of the plant on the map.
-    pub width: i32,
-    /// The height of the plant on the map.
-    pub height: i32,
-    /// The rotation in degrees (0-360) of the plant on the map.
-    pub rotation: f32,
-    /// The x scale of the plant on the map.
-    pub scale_x: f32,
-    /// The y scale of the plant on the map.
-    pub scale_y: f32,
-}
-
-/// The `Planting` entity.
-#[derive(Debug, Clone, AsChangeset)]
-#[diesel(table_name = plantings)]
-pub struct UpdatePlanting {
-    /// The x coordinate of the position on the map.
-    pub x: Option<i32>,
-    /// The y coordinate of the position on the map.
-    pub y: Option<i32>,
-    /// The width of the plant on the map.
-    pub width: Option<i32>,
-    /// The height of the plant on the map.
-    pub height: Option<i32>,
-    /// The rotation of the plant on the map.
-    pub rotation: Option<f32>,
-    /// The x scale of the plant on the map.
-    pub scale_x: Option<f32>,
-    /// The y scale of the plant on the map.
-    pub scale_y: Option<f32>,
 }
